+++ tests/issue_69.rs
@@ -1,5 +1,11 @@
+//! This test must be run under `cargo +nightly miri test` in order to prove
+//! correct handling of the allocation pointer during drop.
+
+#[cfg(feature = "alloc")]
+use bitvec::prelude::*;
+
 #[test]
 #[cfg(feature = "alloc")]
 fn main() {
-	let _: bitvec::boxed::BitBox = bitvec::bitbox![0];
+	let _ = bitbox![0];
 }